+++ src/main/java/com/ista/springboot/web/app/controllers/PeriodoRestController.java
@@ -4,10 +4,11 @@ import com.ista.springboot.web.app.entity.Periodo;
 import com.ista.springboot.web.app.service.IPeriodoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
 
+@CrossOrigin(origins = "*")
 @RestController
 @RequestMapping("/api")
 public class PeriodoRestController {
@@ -21,31 +22,43 @@ public class PeriodoRestController {
     }
 
     @GetMapping("/periodos/{id}")
-    public Periodo show(@PathVariable Integer id) {
-        return periodoService.findById(id);
+    public ResponseEntity<Periodo> show(@PathVariable Integer id) {
+        Periodo periodo = periodoService.findById(id);
+        if (periodo == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(periodo);
     }
 
     @PostMapping("/periodos")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Periodo create(@RequestBody Periodo periodo) {
-        return periodoService.save(periodo);
+    public ResponseEntity<Periodo> create(@RequestBody Periodo periodo) {
+        periodo.setIdPeriodo(null);
+        Periodo nuevoPeriodo = periodoService.save(periodo);
+        return ResponseEntity.status(HttpStatus.CREATED).body(nuevoPeriodo);
     }
 
     @PutMapping("/periodos/{id}")
-    public Periodo update(@RequestBody Periodo periodo, @PathVariable Integer id) {
+    public ResponseEntity<Periodo> update(@RequestBody Periodo periodo, @PathVariable Integer id) {
         Periodo periodoActual = periodoService.findById(id);
+        if (periodoActual == null) {
+            return ResponseEntity.notFound().build();
+        }
         
         periodoActual.setNombre(periodo.getNombre());
         periodoActual.setFechaInicio(periodo.getFechaInicio());
         periodoActual.setFechaFin(periodo.getFechaFin());
         periodoActual.setEstado(periodo.getEstado());
         
-        return periodoService.save(periodoActual);
+        return ResponseEntity.ok(periodoService.save(periodoActual));
     }
 
     @DeleteMapping("/periodos/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+        Periodo periodo = periodoService.findById(id);
+        if (periodo == null) {
+            return ResponseEntity.notFound().build();
+        }
         periodoService.delete(id);
+        return ResponseEntity.noContent().build();
     }
 }
